refactor: centralize asset requirement summaries
